Get configured Consul port as an integer value
This function will fall back to the default if environment variable is set to a blank value.
The assumption here is that a blank value cannot be parsed to an integer in a sensible way and
thus using the default is the best we can do.

Part of rabbitmq/rabbitmq-peer-discovery-consul#41.
Environment variable support is expected to be removed
in the future: rabbitmq/rabbitmq-peer-discovery-consul#40.

See rabbitmq/discussions#104 for the background.